Mutiple pool support: a network can now be sliced into "pools" of APs, clients, and SSIDs.
A pool is a set of physical APs, that announce a set of SSIDs, and
are orchestrated by a set of applications. No two pools can announce
the same SSID. Clients are mapped to a particular pool depending
on the SSID they associate with. Applications are bound to a
particular pool, and can only see clients and APs that are
part of that pool. A typical use case for pools would be to run
two different kinds of mobility managers for an outdoor and
indoor network.